Add baby tracking views and calendar
